:recycle: Split PostgreSQL account read model into directory module by row and tests
Pure mechanical move of driver/src/database/postgres/account.rs into a directory module; no behavior change.
