+++ driver/src/database/postgres/account/mod.rs
@@ -1,90 +1,18 @@
+mod row;
+
+pub use row::PostgresAccountReadModel;
+use row::{account_from_row, AccountRow};
+
 use crate::database::{PostgresConnection, PostgresDatabase};
 use crate::ConvertError;
 use error_stack::Report;
 use kernel::interfaces::read_model::{AccountReadModel, DependOnAccountReadModel};
 use kernel::prelude::entity::{
-    Account, AccountId, AccountIsBot, AccountName, AccountStatus, AuthAccountId, CreatedAt,
-    DeletedAt, EventVersion, Nanoid,
+    Account, AccountId, AccountName, AccountStatus, AuthAccountId, Nanoid,
 };
 use kernel::KernelError;
 use sqlx::types::time::OffsetDateTime;
 use sqlx::PgConnection;
-
-#[derive(sqlx::FromRow)]
-struct AccountRow {
-    id: i64,
-    name: String,
-    is_bot: bool,
-    deleted_at: Option<OffsetDateTime>,
-    version: i64,
-    nanoid: String,
-    created_at: OffsetDateTime,
-    suspended_at: Option<OffsetDateTime>,
-    suspend_expires_at: Option<OffsetDateTime>,
-    suspend_reason: Option<String>,
-    banned_at: Option<OffsetDateTime>,
-    ban_reason: Option<String>,
-}
-
-/// Convert an AccountRow into an Account.
-///
-/// When `check_suspend_expiry` is `false` (used for filtered queries where SQL already
-/// excludes expired suspensions), suspended_at is trusted as-is.
-/// When `true` (used for unfiltered queries), Rust-side expiry check is performed.
-fn account_from_row(value: AccountRow, check_suspend_expiry: bool) -> Account {
-    let status = if let (Some(banned_at), Some(reason)) = (value.banned_at, value.ban_reason) {
-        AccountStatus::Banned { reason, banned_at }
-    } else if let (Some(suspended_at), Some(reason)) =
-        (value.suspended_at, value.suspend_reason.clone())
-    {
-        if check_suspend_expiry {
-            if let Some(expires_at) = value.suspend_expires_at {
-                if expires_at <= OffsetDateTime::now_utc() {
-                    AccountStatus::Active
-                } else {
-                    AccountStatus::Suspended {
-                        reason,
-                        suspended_at,
-                        expires_at: Some(expires_at),
-                    }
-                }
-            } else {
-                AccountStatus::Suspended {
-                    reason,
-                    suspended_at,
-                    expires_at: None,
-                }
-            }
-        } else {
-            AccountStatus::Suspended {
-                reason,
-                suspended_at,
-                expires_at: value.suspend_expires_at,
-            }
-        }
-    } else {
-        AccountStatus::Active
-    };
-
-    Account::new(
-        AccountId::new(value.id),
-        AccountName::new(value.name),
-        AccountIsBot::new(value.is_bot),
-        status,
-        value.deleted_at.map(DeletedAt::new),
-        EventVersion::new(value.version),
-        Nanoid::new(value.nanoid),
-        CreatedAt::new(value.created_at),
-    )
-}
-
-impl From<AccountRow> for Account {
-    fn from(value: AccountRow) -> Self {
-        account_from_row(value, false)
-    }
-}
-
-pub struct PostgresAccountReadModel;
 
 impl AccountReadModel for PostgresAccountReadModel {
     type Executor = PostgresConnection;
@@ -510,275 +438,5 @@ impl DependOnAccountReadModel for PostgresDatabase {
 }
 
 #[cfg(test)]
-mod test {
-    mod read_model {
-        use crate::database::PostgresDatabase;
-        use kernel::interfaces::database::DatabaseConnection;
-        use kernel::interfaces::read_model::{
-            AccountReadModel, AuthAccountReadModel, DependOnAccountReadModel,
-            DependOnAuthAccountReadModel,
-        };
-        use kernel::interfaces::repository::{AuthHostRepository, DependOnAuthHostRepository};
-        use kernel::prelude::entity::{
-            Account, AccountId, AuthAccountId, AuthHostId, DeletedAt, Nanoid,
-        };
-        use kernel::test_utils::{
-            unique_account_name, AccountBuilder, AuthAccountBuilder, AuthHostBuilder,
-        };
-        use sqlx::types::time::OffsetDateTime;
-
-        #[test_with::env(DATABASE_URL)]
-        #[tokio::test]
-        async fn find_by_id() {
-            kernel::ensure_generator_initialized();
-            let database = PostgresDatabase::new().await.unwrap();
-            let mut transaction = database.get_executor().await.unwrap();
-
-            let id = AccountId::default();
-            let account = AccountBuilder::new().id(id.clone()).build();
-            database
-                .account_read_model()
-                .create(&mut transaction, &account)
-                .await
-                .unwrap();
-            let result = database
-                .account_read_model()
-                .find_by_id(&mut transaction, &id)
-                .await
-                .unwrap();
-            assert_eq!(result.as_ref().map(Account::id), Some(account.id()));
-        }
-
-        #[test_with::env(DATABASE_URL)]
-        #[tokio::test]
-        async fn find_by_auth_id() {
-            kernel::ensure_generator_initialized();
-            let database = PostgresDatabase::new().await.unwrap();
-            let mut transaction = database.get_executor().await.unwrap();
-
-            let accounts = database
-                .account_read_model()
-                .find_by_auth_id(&mut transaction, &AuthAccountId::default())
-                .await
-                .unwrap();
-            assert!(accounts.is_empty());
-        }
-
-        #[test_with::env(DATABASE_URL)]
-        #[tokio::test]
-        async fn find_by_auth_id_after_create_and_link() {
-            kernel::ensure_generator_initialized();
-            let database = PostgresDatabase::new().await.unwrap();
-            let mut transaction = database.get_executor().await.unwrap();
-
-            let host_id = AuthHostId::default();
-            let auth_host = AuthHostBuilder::new().id(host_id.clone()).build();
-            database
-                .auth_host_repository()
-                .create(&mut transaction, &auth_host)
-                .await
-                .unwrap();
-
-            let auth_account_id = AuthAccountId::default();
-            let auth_account = AuthAccountBuilder::new()
-                .id(auth_account_id.clone())
-                .host(host_id)
-                .build();
-            database
-                .auth_account_read_model()
-                .create(&mut transaction, &auth_account)
-                .await
-                .unwrap();
-
-            let account = AccountBuilder::new().build();
-            database
-                .account_read_model()
-                .create(&mut transaction, &account)
-                .await
-                .unwrap();
-
-            database
-                .account_read_model()
-                .link_auth_account(&mut transaction, account.id(), &auth_account_id)
-                .await
-                .unwrap();
-
-            let accounts = database
-                .account_read_model()
-                .find_by_auth_id(&mut transaction, &auth_account_id)
-                .await
-                .unwrap();
-            assert_eq!(accounts.len(), 1);
-            assert_eq!(accounts[0].id(), account.id());
-
-            database
-                .account_read_model()
-                .unlink_all_auth_accounts(&mut transaction, account.id())
-                .await
-                .unwrap();
-            database
-                .account_read_model()
-                .deactivate(&mut transaction, account.id())
-                .await
-                .unwrap();
-        }
-
-        #[test_with::env(DATABASE_URL)]
-        #[tokio::test]
-        async fn find_by_name() {
-            kernel::ensure_generator_initialized();
-            let database = PostgresDatabase::new().await.unwrap();
-            let mut transaction = database.get_executor().await.unwrap();
-
-            let name = unique_account_name();
-            let account = AccountBuilder::new().name(name.as_ref()).build();
-            database
-                .account_read_model()
-                .create(&mut transaction, &account)
-                .await
-                .unwrap();
-
-            let result = database
-                .account_read_model()
-                .find_by_name(&mut transaction, &name)
-                .await
-                .unwrap();
-            assert_eq!(result.as_ref().map(Account::id), Some(account.id()));
-            database
-                .account_read_model()
-                .deactivate(&mut transaction, account.id())
-                .await
-                .unwrap();
-        }
-
-        #[test_with::env(DATABASE_URL)]
-        #[tokio::test]
-        async fn find_by_nanoid() {
-            kernel::ensure_generator_initialized();
-            let database = PostgresDatabase::new().await.unwrap();
-            let mut transaction = database.get_executor().await.unwrap();
-
-            let nanoid = Nanoid::default();
-            let account = AccountBuilder::new().nanoid(nanoid.clone()).build();
-            database
-                .account_read_model()
-                .create(&mut transaction, &account)
-                .await
-                .unwrap();
-
-            let result = database
-                .account_read_model()
-                .find_by_nanoid(&mut transaction, &nanoid)
-                .await
-                .unwrap();
-            assert_eq!(result.as_ref().map(Account::id), Some(account.id()));
-            database
-                .account_read_model()
-                .deactivate(&mut transaction, account.id())
-                .await
-                .unwrap();
-        }
-
-        #[test_with::env(DATABASE_URL)]
-        #[tokio::test]
-        async fn create() {
-            kernel::ensure_generator_initialized();
-            let database = PostgresDatabase::new().await.unwrap();
-            let mut transaction = database.get_executor().await.unwrap();
-
-            let account = AccountBuilder::new().build();
-            database
-                .account_read_model()
-                .create(&mut transaction, &account)
-                .await
-                .unwrap();
-            let result = database
-                .account_read_model()
-                .find_by_id(&mut transaction, account.id())
-                .await
-                .unwrap()
-                .unwrap();
-            assert_eq!(result.id(), account.id());
-        }
-
-        #[test_with::env(DATABASE_URL)]
-        #[tokio::test]
-        async fn update() {
-            kernel::ensure_generator_initialized();
-            let database = PostgresDatabase::new().await.unwrap();
-            let mut transaction = database.get_executor().await.unwrap();
-
-            let account = AccountBuilder::new().build();
-            database
-                .account_read_model()
-                .create(&mut transaction, &account)
-                .await
-                .unwrap();
-            let updated_account = AccountBuilder::new()
-                .id(account.id().clone())
-                .name(unique_account_name())
-                .is_bot(true)
-                .build();
-            database
-                .account_read_model()
-                .update(&mut transaction, &updated_account)
-                .await
-                .unwrap();
-            let result = database
-                .account_read_model()
-                .find_by_id(&mut transaction, account.id())
-                .await
-                .unwrap();
-            assert_eq!(result.as_ref().map(Account::id), Some(updated_account.id()));
-        }
-
-        #[test_with::env(DATABASE_URL)]
-        #[tokio::test]
-        async fn deactivate() {
-            kernel::ensure_generator_initialized();
-            let database = PostgresDatabase::new().await.unwrap();
-            let mut transaction = database.get_executor().await.unwrap();
-
-            let account = AccountBuilder::new().build();
-            database
-                .account_read_model()
-                .create(&mut transaction, &account)
-                .await
-                .unwrap();
-
-            database
-                .account_read_model()
-                .deactivate(&mut transaction, account.id())
-                .await
-                .unwrap();
-            let result = database
-                .account_read_model()
-                .find_by_id(&mut transaction, account.id())
-                .await
-                .unwrap();
-            assert!(result.is_none());
-
-            // Ignore if the account is already deleted
-            let account = AccountBuilder::new()
-                .deleted_at(Some(DeletedAt::new(OffsetDateTime::now_utc())))
-                .build();
-            database
-                .account_read_model()
-                .create(&mut transaction, &account)
-                .await
-                .unwrap();
-
-            database
-                .account_read_model()
-                .deactivate(&mut transaction, account.id())
-                .await
-                .unwrap();
-            let result = database
-                .account_read_model()
-                .find_by_id(&mut transaction, account.id())
-                .await
-                .unwrap();
-            assert!(result.is_none());
-        }
-    }
-}
+#[path = "tests.rs"]
+mod test;
